+++ packages/core/api.ts
@@ -1,3 +1,0 @@
-import { transformCode, TransformOptions, TransformOutput } from './index'
-
-export { transformCode, TransformOptions, TransformOutput }