+++ packages/core/src/lib.rs
@@ -4,7 +4,7 @@ use std::time::Instant;
 
 use napi::bindgen_prelude::{Env, Function};
 
-use crate::transform::{TransformOptions, TransformOutput};
+use crate::transform::{CSSData, TransformOptions, TransformOutput};
 
 #[macro_use]
 extern crate napi_derive;
@@ -15,6 +15,7 @@ pub mod parse_css;
 pub mod style_tag;
 pub mod transform;
 pub mod update_attribute;
+pub mod utils;
 
 pub use crate::logger::{log_error, log_info, log_warn};
 
@@ -24,8 +25,8 @@ pub fn transform_code(
   code: String,
   file_path: String,
   options: TransformOptions,
-  css_preprocessor: Option<Function<String, String>>,
-) -> Option<TransformOutput> {
+  css_preprocessor: Option<Function<CSSData, String>>,
+) -> TransformOutput {
   let time = Instant::now();
 
   // Example of using the logging system
@@ -33,13 +34,6 @@ pub fn transform_code(
     println!("Starting transformation for file: {}", file_path);
   }
 
-  let options = transform::TransformOptions {
-    css_out_dir: options.css_out_dir,
-    class_name_list: options.class_name_list,
-    use_theme: options.use_theme,
-    append_timestamp_to_css_file: options.append_timestamp_to_css_file,
-    theme: options.theme,
-  };
   let result = transform::transform(code, file_path, options, css_preprocessor, Some(env));
   let duration = time.elapsed();